+++ dnsext-dox/DNS/DoX.hs
@@ -1,0 +1,14 @@
+module DNS.DoX (
+    dot
+  , doq
+  , doh
+  , doh3
+  , mew
+  , iij
+  ) where
+
+import DNS.DoX.Common
+import DNS.DoX.HTTP2
+import DNS.DoX.HTTP3
+import DNS.DoX.QUIC
+import DNS.DoX.TLS
